requirements.txt
astroid==2.1.0
Click==7.0
Flask==1.0.2
Flask-Script==2.0.6
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pylint==2.2.2
python-dotenv==0.10.0
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
